refactor: rename controllers folder
